moved pacer to separate file, added readme example
